rigolo

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: joterret <joterret@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/26 18:52:11 by jo                #+#    #+#             */
-/*   Updated: 2022/10/27 00:43:57 by joterret         ###   ########.fr       */
+/*   Updated: 2022/12/28 17:06:52 by joterret         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != 0)
